Made it so that you can only remove yourself from cohorts

+++ app/controllers/memberships_controller.rb
@@ -21,8 +21,12 @@ class MembershipsController < ApplicationController
 
   def destroy
     membership = Membership.find(params[:id])
-    membership.destroy
-    redirect_to :back
+    if (current_user.id == membership.user_id)
+      membership.destroy
+      redirect_to :back
+    else
+      redirect_to :back
+    end
   end
 
   private
